[DDW-1152] Cleaning up both Languages and General stories.
